Add Deploy stage

+++ Jenkinsfile
@@ -30,23 +30,34 @@ node {
             }
         }
         stage('Deploy') {
-            try {
-            // Build JAR
-                sh 'mvn clean package -DskipTests'
-                sh '''
-                JAR_FILE=$(ls target/*.jar | grep -v original | head -n 1)
-                nohup java -jar $JAR_FILE --server.port=4000 > app.log 2>&1 &
-                echo $! > app.pid
-                '''
-                input message: 'Aplikasi Java sudah berjalan. Klik "Proceed" untuk menghentikan.'
-                sh '''
-                kill $(cat app.pid)
-                rm app.pid
-                '''
-            } catch (exc) {
-                echo 'Deploy failed!'
-                throw exc
-            }
+        try {
+            // Build
+            sh 'mvn clean package -DskipTests'
+
+            // Stop jika masih ada yang jalan
+            sh 'fuser -k 4000/tcp || true'
+
+            // Run JAR
+            sh '''
+            JAR_FILE=$(find target -name "*.jar" ! -name "original-*.jar" | head -n 1)
+            nohup java -jar $JAR_FILE --server.port=4000 > app.log 2>&1 &
+            echo $! > app.pid
+            '''
+
+            input message: 'Aplikasi Java sudah berjalan. Klik "Proceed" untuk menghentikan.'
+
+            // Stop App
+            sh '''
+            if [ -f app.pid ]; then
+            kill $(cat app.pid) || true
+            rm app.pid
+            fi
+            '''
+
+        } catch (exc) {
+            echo 'Deploy failed!'
+            throw exc
         }
+
     }
 }
